feat: model crud and product crud

+++ backend/ven_inv_service/src/services/productService.ts
@@ -1,4 +1,5 @@
 import { Product } from "../entities/productEntity";
+import { AppError } from "../errors/appError";
 import { ProductRepository } from "../repositories/productRepository";
 
 
@@ -13,11 +14,15 @@ export class ProductService {
         return this.productRepo.getAllProducts();
     }
 
-    async updateProduct(id: number, data: Partial<Product>) {
+    async updateProduct(id: string, data: Partial<Product>) {
         return this.productRepo.updateProduct(id, data);
     }
 
-    async deleteProduct(id: number) {
+    async deleteProduct(id: string) {
+        const product = this.productRepo.findOne(id);
+        if (!product) {
+            throw new AppError("Product not found",404);
+        }
         return this.productRepo.deleteProduct(id);
     }   
 }
